// front/src/app/utils/enums.ts
export enum ThemeColor {
  LIGHT = 'light',
  DARK = 'dark',
}

export enum StatusColor {
  Success = 'success',
  Info = 'info',
  Warning = 'warning',
  Danger = 'danger',
}

export enum FunctionName {
  AddPauser = 'AddPauser',
  AddVerified = 'AddVerified',
  Allowance = 'Allowance',
  Approve = 'Approve',
  ApproveAndCall = 'ApproveAndCall',
  AuthorizeOperator = 'AuthorizeOperator',
  BalanceOf = 'BalanceOf',
  BalanceOfID = 'BalanceOfID',
  BalanceOfBatch = 'BalanceOfBatch',
  Burn = 'Burn',
  BurnData = 'BurnData',
  BurnFrom = 'BurnFrom',
  CancelAndReissue = 'CancelAndReissue',
  ImplementInterfaceForAddress = 'ImplementInterfaceForAddress',
  Cap = 'Cap',
  ChangeOwner = 'ChangeOwner',
  CIDByHash = 'CIDByHash',
  Decimals = 'Decimals',
  DecreaseAllowance = 'DecreaseAllowance',
  DecreaseAllowanceAndCall = 'DecreaseAllowanceAndCall',
  DecreaseSupply = 'DecreaseSupply',
  DefaultOperators = 'DefaultOperators',
  Deployed = 'Deployed',
  GetApproved = 'GetApproved',
  GetCurrentFor = 'GetCurrentFor',
  Granularity = 'Granularity',
  HasHash = 'HasHash',
  HolderAt = 'HolderAt',
  HolderCount = 'HolderCount',
  IncreaseAllowance = 'IncreaseAllowance',
  IncreaseAllowanceAndCall = 'IncreaseAllowanceAndCall',
  IncreaseSupply = 'IncreaseSupply',
  IsApprovedForAll = 'IsApprovedForAll',
  IsHolder = 'IsHolder',
  IsPauser = 'IsPauser',
  IsOperatorFor = 'IsOperatorFor',
  IsOwner = 'IsOwner',
  IsSuperseded = 'IsSuperseded',
  IsVerified = 'IsVerified',
  Mint = 'Mint',
  MintWithTokenURI = 'MintWithTokenURI',
  Name = 'Name',
  NewWallet = 'NewWallet',
  OnErc721Received = 'OnErc721Received',
  OnErc1155BatchReceived = 'OnErc1155BatchReceived',
  OnErc1155Received = 'OnErc1155Received',
  OperatorBurn = 'OperatorBurn',
  OperatorSend = 'OperatorSend',
  Owner = 'Owner',
  OwnerOf = 'OwnerOf',
  Rate = 'Rate',
  RemoveVerified = 'RemoveVerified',
  RenounceOwnership = 'RenounceOwnership',
  RenouncePauser = 'RenouncePauser',
  RevokeOperator = 'RevokeOperator',
  SafeBatchTransferFrom = 'SafeBatchTransferFrom',
  SafeTransferFrom = 'SafeTransferFrom',
  SafeTransferFromData = 'SafeTransferFromData',
  SafeTransferFromValueData = 'SafeTransferFromValueData',
  Send = 'Send',
  SetApprovalForAll = 'SetApprovalForAll',
  SetRate = 'SetRate',
  SupportsInterface = 'SupportsInterface',
  Symbol = 'Symbol',
  Target = 'Target',
  TokensReceived = 'TokensReceived',
  TokensToSend = 'TokensToSend',
  TokenByIndex = 'TokenByIndex',
  TokenFallback = 'TokenFallback',
  TokenOfOwnerByIndex = 'TokenOfOwnerByIndex',
  TokenUri = 'TokenUri',
  TotalSupply = 'TotalSupply',
  Transfer = 'Transfer',
  TransferData = 'TransferData',
  TransferDataFallback = 'TransferDataFallback',
  TransferAndCall = 'TransferAndCall',
  TransferFrom = 'TransferFrom',
  TransferFromAndCall = 'TransferFromAndCall',
  TransferOwnership = 'TransferOwnership',
  UpdateVerified = 'UpdateVerified',
  URI = 'URI',
  Pause = 'Pause',
  Pin = 'Pin',
  Paused = 'Paused',
  Unpause = 'Unpause',
  Upgrade = 'Upgrade',
  Resume = 'Resume',
  Wallet = 'Wallet',
  Withdraw = 'Withdraw',
}

export enum ErcName {
  AllFunctions = 'AllFunctions',
  Xps20 = 'Xps20',
  Xps20Burnable = 'Xps20Burnable',
  Xps20Capped = 'Xps20Capped',
  Xps20Detailed = 'Xps20Detailed',
  Xps20Mintable = 'Xps20Mintable',
  Xps20Pausable = 'Xps20Pausable',
  Xps165 = 'Xps165',
  Xps721 = 'Xps721',
  Xps721Burnable = 'Xps721Burnable',
  Xps721Receiver = 'Xps721Receiver',
  Xps721Metadata = 'Xps721Metadata',
  Xps721Enumerable = 'Xps721Enumerable',
  Xps721Pausable = 'Xps721Pausable',
  Xps721Mintable = 'Xps721Mintable',
  Xps721MetadataMintable = 'Xps721MetadataMintable',
  Xps721Full = 'Xps721Full',
  Xps820 = 'Xps820',
  Xps1155 = 'Xps1155',
  Xps1155Receiver = 'Xps1155Receiver',
  Xps1155Metadata = 'Xps1155Metadata',
  Xps223 = 'Xps223',
  Xps223Receiver = 'Xps223Receiver',
  Xps621 = 'Xps621',
  Xps777 = 'Xps777',
  Xps777Receiver = 'Xps777Receiver',
  Xps777Sender = 'Xps777Sender',
  Xps827 = 'Xps827',
  Xps884 = 'Xps884',
  Upgradeable = 'Upgradeable',
  Ownable = 'Ownable',
  PauserRole = 'PauserRole',
  XpsFS = 'XpsFS',
}